feat: Render pipelines implemented
For every type of shader there should be a pipeline that manages it and every object that makes use of it. This makes the compilation/linking time significally shorter and less resource intensive. TODO: Make a standardization between shaders and objects so objects with the same functionality, i.e. Use of 3D camera, utilize the same shader