feat!: First person camera
The main idea here is that a camera can be considered a real object in the scene, where the lookat function now points to the forward vector of the camera and it can move like any other entity in the scene. The follow method no longer works as intended and needs to be redone in the future.